+++ CHANGELOG.md
@@ -22,8 +22,11 @@ Changes are grouped as follows
 - Fixes type annotations for Functions API. Adds new `FunctionHandle` type for annotating function handles.
 
 ## [7.78.0] - 2025-07-29
-#### Added
+### Added
 - Add `timezone` as an optional param to the WorkflowScheduledTriggerRule.
+- Support for two-phase syncing of instances. See `sync_mode` and `backfill_sort` on the `NodeResultSet` class.
+### Changed
+- Improved error messages when using query-specific fields in sync, and vice versa.
 
 ## [7.77.3] - 2025-07-28
 ### Added